refactor[api-server]: 링크 패키지 lottery 패키지에서 분리
