+++ src/main/java/com/watermelon/server/event/link/controller/LinkController.java
@@ -1,8 +1,9 @@
-package com.watermelon.server.event.lottery.link.controller;
+package com.watermelon.server.event.link.controller;
 
+import com.watermelon.server.event.link.service.LinkService;
+import com.watermelon.server.event.link.utils.LinkUtils;
 import com.watermelon.server.event.lottery.auth.annotations.Uid;
-import com.watermelon.server.event.lottery.link.dto.MyLinkDto;
-import com.watermelon.server.event.lottery.link.service.LinkService;
+import com.watermelon.server.event.link.dto.MyLinkDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -10,8 +11,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
-
-import static com.watermelon.server.event.lottery.link.utils.LinkUtils.makeUrl;
 
 @RestController
 @RequiredArgsConstructor
@@ -33,7 +32,7 @@ public class LinkController {
         HttpHeaders headers = new HttpHeaders();
 
         // 현재 서버의 URL을 기반으로 새로운 경로를 생성
-        headers.add(HttpHeaders.LOCATION, makeUrl(linkService.getUrl(shortedUri)));
+        headers.add(HttpHeaders.LOCATION, LinkUtils.makeUrl(linkService.getUrl(shortedUri)));
         return new ResponseEntity<>(headers, HttpStatus.FOUND);
     }
