+++ src/main/java/com/watermelon/server/event/link/dto/MyLinkDto.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery.link.dto;
+package com.watermelon.server.event.link.dto;
 
 import lombok.Builder;
 import lombok.Data;
