+++ src/main/java/com/watermelon/server/event/link/exception/LinkNotFoundException.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery.link.exception;
+package com.watermelon.server.event.link.exception;
 
 public class LinkNotFoundException extends RuntimeException{
     public LinkNotFoundException() {
