+++ src/main/java/com/watermelon/server/event/link/utils/LinkUtils.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery.link.utils;
+package com.watermelon.server.event.link.utils;
 
 import java.math.BigInteger;
 import java.nio.ByteBuffer;
