+++ src/test/java/com/watermelon/server/ControllerTest.java
@@ -10,8 +10,8 @@ import com.watermelon.server.event.lottery.dto.response.ResponseLotteryRankDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseLotteryWinnerInfoDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseRewardInfoDto;
-import com.watermelon.server.event.lottery.link.dto.MyLinkDto;
-import com.watermelon.server.event.lottery.link.service.LinkService;
+import com.watermelon.server.event.link.dto.MyLinkDto;
+import com.watermelon.server.event.link.service.LinkService;
 import com.watermelon.server.event.parts.dto.response.ResponseMyPartsListDto;
 import com.watermelon.server.event.parts.dto.response.ResponsePartsDrawDto;
 import com.watermelon.server.event.parts.dto.response.ResponseRemainChanceDto;
