+++ src/test/java/com/watermelon/server/lottery/link/service/LinkServiceImplTest.java
@@ -2,9 +2,9 @@ package com.watermelon.server.lottery.link.service;
 
 import com.watermelon.server.event.lottery.domain.Link;
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
-import com.watermelon.server.event.lottery.link.dto.MyLinkDto;
-import com.watermelon.server.event.lottery.link.repository.LinkRepository;
-import com.watermelon.server.event.lottery.link.service.LinkServiceImpl;
+import com.watermelon.server.event.link.dto.MyLinkDto;
+import com.watermelon.server.event.link.repository.LinkRepository;
+import com.watermelon.server.event.link.service.LinkServiceImpl;
 import com.watermelon.server.event.lottery.service.LotteryService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
